% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{all_available}
\alias{all_available}
\title{List all available ISTAT datasets}
\usage{
all_available()
}
\value{
A tibble with the following columns:
\describe{
\item{df_id}{Dataflow ID}
\item{version}{Dataset version}
\item{df_description}{English description of the dataset}
\item{df_structure_id}{Data structure definition ID}
}
}
\description{
Retrieves a list of all available datasets (dataflows) from the ISTAT API.
}
\examples{
\dontrun{
# Get all available datasets
datasets <- all_available()
head(datasets)
}
}
