% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{get_dimension_description}
\alias{get_dimension_description}
\title{Get description for a dimension from its codelist}
\usage{
get_dimension_description(codelist_id)
}
\arguments{
\item{codelist_id}{Codelist ID}
}
\value{
Character string with the dimension description
}
\description{
Get description for a dimension from its codelist
}
\keyword{internal}
