% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Get dimensions for a data structure}
\usage{
get_dimensions(structure_id)
}
\arguments{
\item{structure_id}{Data structure definition ID}
}
\value{
Named list of dimensions with their metadata
}
\description{
Get dimensions for a data structure
}
\keyword{internal}
