% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_name_by_lang_ns}
\alias{get_name_by_lang_ns}
\title{Parse language-specific name from XML with namespace support}
\usage{
get_name_by_lang_ns(node, lang = "en", ns = NULL)
}
\arguments{
\item{node}{XML node containing Name elements}

\item{lang}{Language code (default: "en")}

\item{ns}{XML namespace object}
}
\value{
Character string with the name in the specified language
}
\description{
Parse language-specific name from XML with namespace support
}
\keyword{internal}
