% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{istat_request}
\alias{istat_request}
\title{Make a request to the ISTAT API}
\usage{
istat_request(path, accept = "application/xml", ...)
}
\arguments{
\item{path}{API endpoint path (without base URL)}

\item{accept}{Accept header value (default: application/xml)}

\item{...}{Additional query parameters}
}
\value{
httr2 response object
}
\description{
Make a request to the ISTAT API
}
\keyword{internal}
