% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{istat_request_csv}
\alias{istat_request_csv}
\title{Make a request and return CSV as tibble}
\usage{
istat_request_csv(path, ...)
}
\arguments{
\item{path}{API endpoint path}

\item{...}{Additional query parameters}
}
\value{
tibble with the data
}
\description{
Make a request and return CSV as tibble
}
\keyword{internal}
