% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{istat_request_xml}
\alias{istat_request_xml}
\title{Make a request and return XML with namespaces}
\usage{
istat_request_xml(path, ...)
}
\arguments{
\item{path}{API endpoint path}

\item{...}{Additional query parameters}
}
\value{
A list with 'xml' (the xml document) and 'ns' (the namespaces)
}
\description{
Make a request and return XML with namespaces
}
\keyword{internal}
