% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{reset_filters}
\alias{reset_filters}
\title{Reset all filters to default (all values)}
\usage{
reset_filters(dataset)
}
\arguments{
\item{dataset}{An istat_dataset object}
}
\value{
The modified istat_dataset object with all filters reset to "."
}
\description{
Reset all filters to default (all values)
}
\examples{
\dontrun{
ds <- istat_dataset("139_176")
ds <- set_filters(ds, FREQ = "M")
ds <- reset_filters(ds)  # All filters back to "."
}
}
