% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_attr_safe}
\alias{xml_attr_safe}
\title{Extract attribute from XML node safely}
\usage{
xml_attr_safe(node, attr, default = NA_character_)
}
\arguments{
\item{node}{XML node}

\item{attr}{Attribute name}

\item{default}{Default value if not found}
}
\value{
Character string
}
\description{
Extract attribute from XML node safely
}
\keyword{internal}
