% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_text_safe}
\alias{xml_text_safe}
\title{Extract text from XML node safely}
\usage{
xml_text_safe(node, xpath, default = NA_character_)
}
\arguments{
\item{node}{XML node}

\item{xpath}{XPath expression}

\item{default}{Default value if not found}
}
\value{
Character string
}
\description{
Extract text from XML node safely
}
\keyword{internal}
